Demonstrate Field with out value property as direct attribute.
This is the simplest way I could think of to demonstrate the issue I am seeing.
An input control of type radio requires the value property to be set in the direct attribute parameters passed to Field. It does not work if the value is passed in via the 'props' property, this is why I think the type definition requires at least "value" be a direct if optional parameter on the Field interface on BaseFieldProps.